+++ app/src/main/java/org/sopt/and/presentation/ui/auth/SignUpScreen.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import org.sopt.and.navigation.AuthNavItem
-import org.sopt.and.presentation.ui.component.CustomTextField
-import org.sopt.and.presentation.ui.component.TextFieldValidateResult
+import org.sopt.and.presentation.ui.auth.component.CustomTextField
+import org.sopt.and.presentation.ui.auth.component.TextFieldValidateResult
 import org.sopt.and.presentation.utils.showToast
 import org.sopt.and.presentation.viewmodel.SignUpViewModel
