+++ app/src/main/java/org/sopt/and/presentation/ui/auth/component/AuthServiceIcon.kt
@@ -1,0 +1,2 @@
+package org.sopt.and.presentation.ui.auth.component
+
