--- app/src/main/java/org/sopt/and/presentation/ui/component/CustomTextFieldDescription.kt
@@ -1,2 +1,0 @@
-package org.sopt.and.presentation.ui.component
-
