+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,8 @@
     <string name="my_content_desc">MY</string>
     <string name="homeview_wave_editor_recommendation">믿고 보는 웨이브 에디터 추천작</string>
     <string name="homeview_today_top_20">오늘의 TOP 20</string>
+
+    <string name="signup_email_description">로그인, 비밀번호 찾기, 알림에 사용되니 정확한 이메일을 입력해주세요.</string>
+    <string name="signup_password_description">비밀번호는 8~20자 이내로 영문 대소문자, 숫자, 특수문자 중 3가지 이상 혼용하여 입력해주세요.</string>
+
 </resources>
